+++ lib/src/material/material_desktop_controls.dart
@@ -1,16 +1,16 @@
 import 'dart:async';
 
-import 'package:chewie/src/animated_play_pause.dart';
-import 'package:chewie/src/center_play_button.dart';
-import 'package:chewie/src/chewie_player.dart';
-import 'package:chewie/src/chewie_progress_colors.dart';
-import 'package:chewie/src/helpers/utils.dart';
-import 'package:chewie/src/material/material_progress_bar.dart';
-import 'package:chewie/src/material/widgets/options_dialog.dart';
-import 'package:chewie/src/material/widgets/playback_speed_dialog.dart';
-import 'package:chewie/src/models/option_item.dart';
-import 'package:chewie/src/models/subtitle_model.dart';
-import 'package:chewie/src/notifiers/index.dart';
+import 'package:chewie_flower/src/animated_play_pause.dart';
+import 'package:chewie_flower/src/center_play_button.dart';
+import 'package:chewie_flower/src/chewie_player.dart';
+import 'package:chewie_flower/src/chewie_progress_colors.dart';
+import 'package:chewie_flower/src/helpers/utils.dart';
+import 'package:chewie_flower/src/material/material_progress_bar.dart';
+import 'package:chewie_flower/src/material/widgets/options_dialog.dart';
+import 'package:chewie_flower/src/material/widgets/playback_speed_dialog.dart';
+import 'package:chewie_flower/src/models/option_item.dart';
+import 'package:chewie_flower/src/models/subtitle_model.dart';
+import 'package:chewie_flower/src/notifiers/index.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
@@ -136,6 +136,7 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
 
   @override
   void dispose() {
+    _chewieController?.removeListener(_onChewieChanged);
     _dispose();
     _focusNode.dispose();
     super.dispose();
@@ -155,11 +156,24 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
     controller = chewieController.videoPlayerController;
 
     if (oldController != chewieController) {
+      oldController?.removeListener(_onChewieChanged);
+      _chewieController?.addListener(_onChewieChanged);
       _dispose();
       _initialize();
     }
 
     super.didChangeDependencies();
+  }
+
+  void _onChewieChanged() {
+    if (!mounted) return;
+    final newController = chewieController.videoPlayerController;
+    if (!identical(controller, newController)) {
+      _dispose();
+      controller = newController;
+      _initialize();
+      if (mounted) setState(() {});
+    }
   }
 
   Widget _buildSubtitleToggle({IconData? icon, bool isPadded = false}) {